# gateway/api/swagger.yaml
basePath: /
definitions:
  auth.AuthUserResponse:
    properties:
      email:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.AuthUserResponse'
    type: object
  auth.ProfileResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      position:
        type: string
      updated_at:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      department:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone:
        type: string
      position:
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      profile:
        $ref: '#/definitions/auth.ProfileResponse'
      token:
        type: string
      user:
        $ref: '#/definitions/auth.AuthUserResponse'
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        example: access denied
        type: string
    type: object
...
  project.CreateProjectEventRequest:
    properties:
      actual_date:
        type: string
      description:
        type: string
      name:
        type: string
      planned_date:
        type: string
      stage_id:
        type: string
      status:
        type: string
    type: object
...
  project.ListProjectEventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/project.ProjectEventResponse'
        type: array
    type: object
...
  project.ProjectEventResponse:
    properties:
      actual_date:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      planned_date:
        type: string
      project_id:
        type: string
      stage_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  project.UpdateProjectEventRequest:
    properties:
      actual_date:
        type: string
      description:
        type: string
      name:
        type: string
      planned_date:
        type: string
      stage_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:28080
info:
  contact: {}
  description: Public REST API exposed by the gateway. Internal microservices communicate
    only over gRPC and are intentionally omitted from this specification.
  title: Project Management Gateway API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password and returns a JWT access
        token.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/auth/me:
    get:
      description: Returns JWT claims of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current auth user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a regular user account and returns a JWT access token.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Register a user
      tags:
      - auth
...
  /api/project-events/{id}:
    delete:
      description: Deletes a project event. Accessible to administrators and the project
        manager.
      parameters:
      - description: Project event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project event
      tags:
      - project-events
    get:
      description: Returns a project event by ID. Accessible to administrators, the
        project manager, and active project members.
      parameters:
      - description: Project event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectEventResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project event
      tags:
      - project-events
    put:
      consumes:
      - application/json
      description: Updates a project event. Accessible to administrators and the project
        manager.
      parameters:
      - description: Project event ID
        in: path
        name: id
        required: true
        type: string
      - description: Project event update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.UpdateProjectEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project event
      tags:
      - project-events
...
  /api/projects/{projectId}/events:
    get:
      description: Returns project events. Accessible to administrators, the project
        manager, and active project members.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListProjectEventsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List project events
      tags:
      - project-events
    post:
      consumes:
      - application/json
      description: Creates a project event. Accessible to administrators and the project
        manager.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Project event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.ProjectEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create project event
      tags:
      - project-events
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT access token. Use the `Bearer {token}` format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"